filterByLogLevels moved to helpers
